+++ src/anomalib/pipelines/hpo/wandb_job.py
@@ -3,57 +3,24 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-import logging
 from tempfile import TemporaryDirectory
 
-from lightning.pytorch.loggers import Logger
-
-from anomalib.data import AnomalibDataModule, get_datamodule
+from anomalib.data import get_datamodule
 from anomalib.engine import Engine
-from anomalib.models import AnomalyModule, get_model
+from anomalib.models import get_model
 from anomalib.pipelines.components import Job
+from anomalib.utils.exceptions import try_import
 from anomalib.utils.logging import hide_output
 
 from .utils import flatten_hpo_params, set_in_nested_dict
 
-logger = logging.getLogger(__name__)
+if try_import("wandb"):
+    import wandb
+    from lightning.pytorch.loggers import WandbLogger
 
-
-class CometHPOJob(Job):
-    """Comet based HPO job."""
-
-    name = "comet_hpo"
-
-    def __init__(self, model: AnomalyModule, datamodule: AnomalibDataModule, logger: Logger) -> None:
-        super().__init__()
-        self.model = model
-        self.datamodule = datamodule
-        self.logger = logger
-
-    @hide_output
-    def run(
-        self,
-        task_id: int | None = None,
-    ) -> None:
-        """Run the HPO."""
-        del task_id  # Not needed for this job
-        with TemporaryDirectory() as temp_dir:
-            engine = Engine(
-                default_root_dir=temp_dir,
-                logger=self.logger,
-            )
-            engine.fit(self.model, datamodule=self.datamodule)
-            self.logger.finalize(status="completed")
-
-    @staticmethod
-    def collect(_: list) -> None:
-        """Does not collect any results."""
-        return
-
-    @staticmethod
-    def save(_: None = None) -> None:
-        """Does not save any results."""
-        return
+    WAND_AVAILABLE = True
+else:
+    WAND_AVAILABLE = False
 
 
 class WandbHPOJob(Job):
@@ -69,8 +36,9 @@ class WandbHPOJob(Job):
     name = "wandb_hpo"
 
     def __init__(self, project: str, entity: str, sweep_configuration: dict, count: int) -> None:
-        import wandb
-
+        if not WAND_AVAILABLE:
+            msg = "HPO using wandb is requested but wandb is not installed."
+            raise ImportError(msg)
         self.original_config = sweep_configuration.copy()
         flattened_hpo_parameters = flatten_hpo_params(sweep_configuration["parameters"])
         sweep_configuration["parameters"] = flattened_hpo_parameters
@@ -81,16 +49,12 @@ class WandbHPOJob(Job):
     @hide_output
     def run(self, task_id: int | None = None) -> None:
         """Run the HPO."""
-        import wandb
-
         del task_id  # Not needed for this job
 
         wandb.agent(self.sweep_id, function=self.train, count=self.count)
 
     def train(self) -> None:
         """Training method that is called by ``wandb.agent``."""
-        from lightning.pytorch.loggers import WandbLogger
-
         wandb_logger = WandbLogger(config=self.sweep_configuration)
         experiment_config = wandb_logger.experiment.config
